Change in EventQuery and the things that also needed due to this change
Now the EventQuery now process the next in the moveNext method instead of in the creation and saving the result in a treeset
EventStorage now saves the events in a LinkedHashSet instead of a simple HashSet to keep the order of the insertions.
Querys are not ordered anymore by ascending order, instead they are orderer by insert order.
Tests changed to match the new behavior of the iterator